+++ demos/main/src/main/java/com/google/android/exoplayer2/demo/DownloadTracker.java
@@ -37,7 +37,7 @@ import com.google.android.exoplayer2.offline.ActionFile;
 import com.google.android.exoplayer2.offline.DownloadAction;
 import com.google.android.exoplayer2.offline.DownloadHelper;
 import com.google.android.exoplayer2.offline.DownloadManager;
-import com.google.android.exoplayer2.offline.DownloadManager.TaskState;
+import com.google.android.exoplayer2.offline.DownloadManager.DownloadState;
 import com.google.android.exoplayer2.offline.DownloadService;
 import com.google.android.exoplayer2.offline.ProgressiveDownloadHelper;
 import com.google.android.exoplayer2.offline.StreamKey;
@@ -144,11 +144,11 @@ public class DownloadTracker implements DownloadManager.Listener {
   }
 
   @Override
-  public void onTaskStateChanged(DownloadManager downloadManager, TaskState taskState) {
-    DownloadAction action = taskState.action;
+  public void onDownloadStateChanged(DownloadManager downloadManager, DownloadState downloadState) {
+    DownloadAction action = downloadState.action;
     Uri uri = action.uri;
-    if ((action.isRemoveAction && taskState.state == TaskState.STATE_COMPLETED)
-        || (!action.isRemoveAction && taskState.state == TaskState.STATE_FAILED)) {
+    if ((action.isRemoveAction && downloadState.state == DownloadState.STATE_COMPLETED)
+        || (!action.isRemoveAction && downloadState.state == DownloadState.STATE_FAILED)) {
       // A download has been removed, or has failed. Stop tracking it.
       if (trackedDownloadStates.remove(uri) != null) {
         handleTrackedDownloadStatesChanged();