+++ RELEASENOTES.md
@@ -27,6 +27,7 @@
   ([#5169](https://github.com/google/ExoPlayer/issues/5169)).
 * DownloadManager:
   * Create only one task for all DownloadActions for the same content.
+  * Rename TaskState to DownloadState.
 
 ### 2.9.2 ###